+++ packages/origin-ui-core/src/components/Modal/index.ts
@@ -2,3 +2,4 @@ export * from './RequestCertificatesModal';
 export * from './AccountMismatchModal';
 export * from './RoleChangedModal';
 export * from './ConnectBlockchainAccountModal';
+export * from './NoBlockchainAccountModal';
