+++ packages/origin-ui-core/src/utils/index.ts
@@ -14,3 +14,4 @@ export * from './clients';
 export { useTranslation } from 'react-i18next';
 export * from './colors';
 export * from './blockchainNetwork';
+export * from './organizationRoles';
